Moved some classes around to break package cycle.
