+++ subprojects/testing-native/src/main/java/org/gradle/nativeplatform/test/xctest/internal/execution/XCTestScraper.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.nativeplatform.test.xctest.internal;
+package org.gradle.nativeplatform.test.xctest.internal.execution;
 
 import com.google.common.base.Joiner;
 import org.gradle.api.internal.tasks.testing.DefaultTestClassDescriptor;
