+++ subprojects/testing-native/src/main/java/org/gradle/nativeplatform/test/xctest/internal/execution/XCTestSelection.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.nativeplatform.test.xctest.internal;
+package org.gradle.nativeplatform.test.xctest.internal.execution;
 
 import com.google.common.collect.ImmutableList;
 import org.apache.commons.lang.StringUtils;
