+++ subprojects/testing-native/src/main/java/org/gradle/nativeplatform/test/xctest/tasks/XCTest.java
@@ -28,9 +28,9 @@ import org.gradle.api.tasks.Internal;
 import org.gradle.api.tasks.Optional;
 import org.gradle.api.tasks.SkipWhenEmpty;
 import org.gradle.api.tasks.testing.AbstractTestTask;
-import org.gradle.nativeplatform.test.xctest.internal.XCTestSelection;
-import org.gradle.nativeplatform.test.xctest.internal.XCTestTestExecutionSpec;
-import org.gradle.nativeplatform.test.xctest.internal.XCTestExecuter;
+import org.gradle.nativeplatform.test.xctest.internal.execution.XCTestSelection;
+import org.gradle.nativeplatform.test.xctest.internal.execution.XCTestTestExecutionSpec;
+import org.gradle.nativeplatform.test.xctest.internal.execution.XCTestExecuter;
 
 import java.io.File;
 import java.util.List;
