+++ subprojects/testing-native/src/test/groovy/org/gradle/nativeplatform/test/xctest/internal/execution/XCTestSelectionTest.groovy
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.nativeplatform.test.xctest.internal
+package org.gradle.nativeplatform.test.xctest.internal.execution
 
 import org.gradle.testing.internal.util.Specification
 import spock.lang.Subject
